moved unit tests to correct file
