+++ test/unit/io/test_nice_serialization.py
@@ -69,56 +69,3 @@ class NiceSerializationTester(BaseCase):
         self.assertTrue(mdl_cloud.is_equivalent(mdl_cloud2))
 
 
-#Somewhat random question for other pyGSTi developers: 
-#Am I missing something or does this class not test anything related to serialization?
-class ModelEquivalenceTester(BaseCase):
-
-    def setUp(self):
-        nQubits = 2
-        self.pspec_2Q = QubitProcessorSpec(nQubits, ('Gx', 'Gy', 'Gcnot'), geometry="line",
-                                           qubit_labels=['qb{}'.format(i) for i in range(nQubits)])
-
-    def check_model(self, mdl):
-        mcopy = mdl.copy()
-        self.assertFalse(mdl is mcopy)
-        self.assertTrue(mcopy.is_similar(mdl))
-        self.assertTrue(mcopy.is_equivalent(mdl))
-
-        if mdl.num_params > 0:
-            r = np.random.random(mdl.num_params)
-            if np.linalg.norm(r) < 1e6:  # just in case we randomly get all zeros!
-                r = 0.1 * np.ones(mcopy.num_params)
-            v_prime = mdl.to_vector() + r
-            mcopy.from_vector(v_prime)
-            self.assertTrue(mcopy.is_similar(mdl))
-            self.assertFalse(mcopy.is_equivalent(mdl))
-
-    def test_explicit_model_equal(self):
-        mdl_explicit = smq1Q_XYI.target_model()
-        self.check_model(mdl_explicit)
-
-    def test_local_model_equal(self):
-        mdl_local = create_crosstalk_free_model(self.pspec_2Q,
-                                                ideal_gate_type='H+S', ideal_spam_type='tensor product H+S',
-                                                independent_gates=False,
-                                                ensure_composed_gates=False)
-        self.check_model(mdl_local)
-
-        mdl_local = create_crosstalk_free_model(self.pspec_2Q,
-                                                ideal_gate_type='H+S', ideal_spam_type='computational',
-                                                independent_gates=True,
-                                                ensure_composed_gates=True)
-        self.check_model(mdl_local)
-
-    def test_cloud_model_equal(self):
-        mdl_cloud = create_cloud_crosstalk_model(self.pspec_2Q, depolarization_strengths={'Gx': 0.05},
-                                                 stochastic_error_probs={'Gy': (0.01, 0.02, 0.03)},
-                                                 lindblad_error_coeffs={'Gcnot': {('H','ZZ'): 0.07, ('S','XX'): 0.10}},
-                                                 independent_gates=False, independent_spam=True, verbosity=2)
-        self.check_model(mdl_cloud)
-
-        mdl_cloud = create_cloud_crosstalk_model(self.pspec_2Q, depolarization_strengths={'Gx': 0.05},
-                                                 stochastic_error_probs={'Gy': (0.01, 0.02, 0.03)},
-                                                 lindblad_error_coeffs={'Gcnot': {('H','ZZ'): 0.07, ('S','XX'): 0.10}},
-                                                 independent_gates=True, independent_spam=True, verbosity=2)
-        self.check_model(mdl_cloud)
