+++ test/unit/objects/test_model.py
@@ -15,6 +15,9 @@ from pygsti.modelmembers.operations import LinearOperator, FullArbitraryOp
 from pygsti.models import ExplicitOpModel
 from pygsti.circuits import Circuit
 from pygsti.models.gaugegroup import FullGaugeGroupElement
+from pygsti.modelpacks import smq1Q_XYI
+from pygsti.processors import QubitProcessorSpec
+from pygsti.models import create_crosstalk_free_model, create_cloud_crosstalk_model
 from ..util import BaseCase, needs_cvxpy
 SKIP_DIAMONDIST_ON_WIN = True
 
@@ -210,7 +213,7 @@ class GeneralMethodBase(object):
 
     @needs_cvxpy
     def test_diamonddist(self):
-        if SKIP_DIAMONDIST_ON_WIN and sys.platform.startswith('win'): return
+        if  SKIP_DIAMONDIST_ON_WIN and sys.platform.startswith('win'): return
         cp = self.model.copy()
         self.assertAlmostEqual(self.model.diamonddist(cp), 0)
         # TODO non-trivial case
@@ -743,3 +746,55 @@ class FullHighThresholdMethodTester(FullModelBase, ThresholdMethodBase, BaseCase
 #    def test_randomize_with_unitary_raises(self):
 #        with self.assertRaises(AssertionError):
 #            self.model.randomize_with_unitary(1, rand_state=np.random.RandomState())  # scale shouldn't matter
+
+class ModelEquivalenceTester(BaseCase):
+
+    def setUp(self):
+        nQubits = 2
+        self.pspec_2Q = QubitProcessorSpec(nQubits, ('Gx', 'Gy', 'Gcnot'), geometry="line",
+                                           qubit_labels=['qb{}'.format(i) for i in range(nQubits)])
+
+    def check_model(self, mdl):
+        mcopy = mdl.copy()
+        self.assertFalse(mdl is mcopy)
+        self.assertTrue(mcopy.is_similar(mdl))
+        self.assertTrue(mcopy.is_equivalent(mdl))
+
+        if mdl.num_params > 0:
+            r = np.random.random(mdl.num_params)
+            if np.linalg.norm(r) < 1e6:  # just in case we randomly get all zeros!
+                r = 0.1 * np.ones(mcopy.num_params)
+            v_prime = mdl.to_vector() + r
+            mcopy.from_vector(v_prime)
+            self.assertTrue(mcopy.is_similar(mdl))
+            self.assertFalse(mcopy.is_equivalent(mdl))
+
+    def test_explicit_model_equal(self):
+        mdl_explicit = smq1Q_XYI.target_model()
+        self.check_model(mdl_explicit)
+
+    def test_local_model_equal(self):
+        mdl_local = create_crosstalk_free_model(self.pspec_2Q,
+                                                ideal_gate_type='H+S', ideal_spam_type='tensor product H+S',
+                                                independent_gates=False,
+                                                ensure_composed_gates=False)
+        self.check_model(mdl_local)
+
+        mdl_local = create_crosstalk_free_model(self.pspec_2Q,
+                                                ideal_gate_type='H+S', ideal_spam_type='computational',
+                                                independent_gates=True,
+                                                ensure_composed_gates=True)
+        self.check_model(mdl_local)
+
+    def test_cloud_model_equal(self):
+        mdl_cloud = create_cloud_crosstalk_model(self.pspec_2Q, depolarization_strengths={'Gx': 0.05},
+                                                 stochastic_error_probs={'Gy': (0.01, 0.02, 0.03)},
+                                                 lindblad_error_coeffs={'Gcnot': {('H','ZZ'): 0.07, ('S','XX'): 0.10}},
+                                                 independent_gates=False, independent_spam=True, verbosity=2)
+        self.check_model(mdl_cloud)
+
+        mdl_cloud = create_cloud_crosstalk_model(self.pspec_2Q, depolarization_strengths={'Gx': 0.05},
+                                                 stochastic_error_probs={'Gy': (0.01, 0.02, 0.03)},
+                                                 lindblad_error_coeffs={'Gcnot': {('H','ZZ'): 0.07, ('S','XX'): 0.10}},
+                                                 independent_gates=True, independent_spam=True, verbosity=2)
+        self.check_model(mdl_cloud)
